Nexthikes-project-10-/requirements.txt
streamlit
opencv-python
pytesseract
reportlab
Pillow
numpy
onnxruntime
